change if loader logic to opacity
